# app/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "description": "Code With Prince",
      "version": "4.11.1",
      "title": "Code With Prince velaiapp"
     
    },
...
      "servers": [
        {
        "url": "/"
        },{
            "url":"/http"
        }
    ],
...
    "components": {
        "schemas": {
            "TodoList": {
                "type": "object",
                "properties": {
                    "number": {
                    "type": "string"
                    }
                   
                }
            },
            "user_id":{
                "type" : "integer",
                "properties" : {
                    "user_id" : {
                        "type" : "integer"
                    }
                }
            },
            "update":{
                "type": "object",
                "properties" :{
                  "username": {
                      "type": "string"
                      },"emailid":{
                        "type":"string"
                      },"dob":{"type":"string"},"gender":{"type":"string"},"additionalnumber":{
                        "type":"string"
                      },"proof":{"type":"string"},"resume":{"type":"string"},
                      "profilepic":{"type":"string"
                    }
                      
                }
            }
        }
    },
...
    "paths": {
...
            "/api/user/education": {
                "post": {
                "tags": [
                    "post  API"
                ],
                "summary": "Add edu",
                "requestBody": {
                "description": "Post New todo To Database",
                "required": true,
                "content": {
                    "application/json": {
                    "schema": {
                        "$ref": "#/components2/schemas/TodoList"
                        }
                    }
                }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                        "$ref": "#/components/schemas/TodoList"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                        "$ref": "#/components/schemas/TodoList"
                        }
                    }
                }
        
                }
            
            
            },
                "/api/user/experience": {
                    "post": {
                    "tags": [
                        "post  API"
                    ],
                    "summary": "Add exp",
                    "requestBody": {
                    "description": "Post New todo To Database",
                    "required": true,
                    "content": {
                        "application/json": {
                        "schema": {
                            "$ref": "#/components3/schemas/TodoList"
                            }
                        }
                    }
                    },
                    "responses": {
                        "200": {
                            "description": "OK",
                            "schema": {
                            "$ref": "#/components/schemas/TodoList"
                            }
                        },
                        "404": {
                            "description": "NOT FOUND",
                            "schema": {
                            "$ref": "#/components/schemas/TodoList"
                            }
                        }
                    }
            
                }
                
                
            },
...
                "/api/rentsee_choice": {
                    "post": {
                    "tags": [
                        " Rental post API"
                    ],
                    "summary": "rental seeker choice ",
                    "requestBody": {
                    "description": "Post New todo To Database",

                    "required": true,
                   
                    "content": {
                        
                        "application/json": {
                        "schema": {
                            "$ref": "#/rentalsee1/schemas/TodoList"
                            }
                        }
                    }
                    },
                    "responses": {
                        "200": {
                            "description": "OK",
                            "schema": {
                            "$ref": "#/components/schemas/TodoList"
                            }
                        },
                        "404": {
                            "description": "NOT FOUND",
                            "schema": {
                            "$ref": "#/components/schemas/TodoList"
                            }
                        }
                    }

                }
                
                
            },
            "/api/rental_see_userinfo_details": {
                "post": {
                "tags": [
                    " Rental post API"
                ],
                "summary": "rental seeker info ",
                "requestBody": {
                "description": "Post New todo To Database",

                "required": true,
               
                "content": {
                    
                    "application/json": {
                    "schema": {
                        "$ref": "#/rentalsee2/schemas/TodoList"
                        }
                    }
                }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                        "$ref": "#/components/schemas/TodoList"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                        "$ref": "#/components/schemas/TodoList"
                        }
                    }
                }

            }
            
            
        },
        "/api/rental_call_history": {
            "post": {
            "tags": [
                " Rental post API"
            ],
            "summary": "rental seeker info ",
            "requestBody": {
            "description": "Post New todo To Database",

            "required": true,
           
            "content": {
                
                "application/json": {
                "schema": {
                    "$ref": "#/rentalsee3/schemas/TodoList"
                    }
                }
            }
            },
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                    "$ref": "#/components/schemas/TodoList"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                    "$ref": "#/components/schemas/TodoList"
                    }
                }
            }

        }
        
        
    },
...
                    "/api/rent_post_fillter/{product_type}":{

                        "get" :{
                            "tags": [
                                "rental Get API"
                            ],
                                "summary": "Get ",
                                "parameters": [{
                                "name": "product_type",
                                "in": "path",
                                "description": "job_apply_count",
                                "required": true,
                                "type": "string",
                                "schema" : {
                                    "$ref": "#/rentalsee1/schemas/product_type"
                                }
                                }],
                            
                                "responses": {
                                "200": {
                                    "description": "OK",
                                    "schema": {
                                    "$ref": "#/components/schemas/TodoList"
                                    }
                                },
                                "404": {
                                    "description": "NOT FOUND",
                                    "schema": {
                                    "$ref": "#/components/schemas/TodoList"
                                    }
                                }
                            }
                        }
                    },
                    "/api/rental_see_call_history/{user_id}":{

                        "get" :{
                            "tags": [
                                "rental Get API"
                            ],
                            "summary": "Get ",
                            "parameters": [{
                            "name": "user_id",
                            "in": "path",
                            "description": "job_apply_count",
                            "required": true,
                            "type": "string",
                            "schema" : {
                                "$ref": "#/rentalsee2/schemas/user_id"
                            }
                            }],
                        
                            "responses": {
                            "200": {
                                "description": "OK",
                                "schema": {
                                "$ref": "#/components/schemas/TodoList"
                                }
                            },
                            "404": {
                                "description": "NOT FOUND",
                                "schema": {
                                "$ref": "#/components/schemas/TodoList"
                                }
                            }
                        }
                    }
                },
...
                "/api/user/{user_id}":{
                    "put": {
                        "tags": [
                            "Update All  API"
                        ],
                        "parameters": [{
                            "name": "user_id",
                            "in": "path",
                            "description": "todo id to update",
                            "required": true,
                            "type": "integer",
                            "schema" : {
                                "$ref": "#/components/schemas/user_id"
                            }
                        }],
                        "summary": "Update Todo",
                        "requestBody": {
                        "description": "Update Existing todo On Database",
                        "required": true,
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/update"
                                }
                            }
                        }
                        },
                        "responses": {
                            "200": {
                                "description": "OK",
                                "schema": {
                                "$ref": "#/components/schemas/TodoList"
                                }
                            },
                            "404": {
                                "description": "NOT FOUND",
                                "schema": {
                                "$ref": "#/components/schemas/TodoList"
                                }
                            }
                        }
                    }
                
                }  ,
                "/api/user_details_update/{user_id}":{
                    "put": {
                        "tags": [
                            "Update All  API"
                        ],
                        "parameters": [{
                            "name": "user_id",
                            "in": "path",
                            "description": "all type user profile details update",
                            "required": true,
                            "type": "integer",
                            "schema" : {
                                "$ref": "#/components1/schemas/user_id"
                            }
                        }],
                        "summary": "Update Todo",
                        "requestBody": {
                        "description": "Update Existing todo On Database",
                        "required": true,
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components1/schemas/update"
                                }
                            }
                        }
                        },
                        "responses": {
                            "200": {
                                "description": "OK",
                                "schema": {
                                "$ref": "#/components/schemas/TodoList"
                                }
                            },
                            "404": {
                                "description": "NOT FOUND",
                                "schema": {
                                "$ref": "#/components/schemas/TodoList"
                                }
                            }
                        }
                    }
                
                }    
    }
}